requirements.txt
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
Faker==36.2.2
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Limiter==3.10.1
gevent==24.11.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
polars==1.24.0
pyarrow==19.0.1
Pygments==2.19.1
pyrate-limiter==2.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-cache==1.2.1
requests-limiter==0.0.1
requests-ratelimiter==0.7.0
rich==13.9.4
setuptools==75.8.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
url-normalize==1.4.3
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
